agent_service/src/requirements.txt
langchain
openai
pinecone-client
python-dotenv
tiktoken
